Bump simplejson from 3.8.2 to 3.16.1
Simplejson is under active development. There have been a lot of bugfixes and speed upgrades. I just want this more
recent version!

Askbot uses precisely loads() and dumps() which are still present in the recent version and their signatures did not
change. More interessting is the fact that some form of "patch" is applied to django to (re?)enable simplejson. This has
to be investigated again when changing django versions. Simplejson is also a very popular project. I will simply do the
upgrade for now.

NOTE: revisit with django upgrades